% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrap_dm.R
\name{camtrap_dm}
\alias{camtrap_dm}
\title{camtrap_dm}
\usage{
camtrap_dm(
  ctTable,
  gridId,
  camId,
  speCol,
  speName,
  dateCol,
  Xcol,
  Ycol,
  numOcc
)
}
\arguments{
\item{ctTable}{A dataframe from camera trap survey.}

\item{gridId}{A column contain grid ID.}

\item{camId}{A column contain camera ID.}

\item{speCol}{A column containing all the species for which to extract detection matrices.}

\item{speName}{A selected species name within the "speCol" column to be extracted as detection matrices.}

\item{dateCol}{A column contain date information (in POSIXct format).}

\item{Xcol}{A quoted column name representing X coordinates within the dataFrame.}

\item{Ycol}{A quoted column name representing Y coordinates within the dataFrame.}

\item{numOcc}{Number of days or times used for an occation.}
}
\value{
A data frame containing the detection matrix for selected species from camera trap survey
}
\description{
A function to extract detection matrices for a species from camtrap survey.
}
