% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFragstat.R
\name{gridFragstat}
\alias{gridFragstat}
\title{Calculate landscape fragmentation matrices within grid cells}
\usage{
gridFragstat(
  landCover,
  landLevel = "landscape",
  lClass = NULL,
  subGrids,
  lsmFunc,
  as_Raster = TRUE,
  rasRes = 1000,
  export_path = NULL
)
}
\arguments{
\item{landCover}{A SpatRaster (terra) object of land cover classes in UTM projection.}

\item{landLevel}{Character, either "landscape" or "class" level.}

\item{lClass}{Numeric or logical, class value for extraction (only used if landLevel = "class").}

\item{subGrids}{A SpatVector or sf object of grid cells (must contain unique 'id').}

\item{lsmFunc}{Function from \code{landscapemetrics} package (e.g., \code{lsm_l_area_mn}).}

\item{as_Raster}{Logical, if TRUE outputs rasterized grid results.}

\item{rasRes}{Numeric, raster resolution in meters if as_Raster = TRUE.}

\item{export_path}{Optional file path to export GeoTIFF (set NAflag = -999 automatically).}
}
\value{
A SpatRaster (if as_Raster = TRUE) or sf object (if FALSE).
}
\description{
Calculates landscape fragmentation metrics (using landscapemetrics functions)
within predefined grid cells. Supports landscape- or class-level metrics,
and outputs either polygons or raster.
}
\examples{
# gridFragstat(landCover, "landscape", subGrids, lsm_l_area_mn)

}
