% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc_changes.R
\name{lulc_changes}
\alias{lulc_changes}
\title{lulc_changes}
\usage{
lulc_changes(dataFrame, lcYear1, lcYear2, type = c("long", "short"))
}
\arguments{
\item{dataFrame}{Data frame containing land cover data.}

\item{lcYear1}{Column name (string) for year 1.}

\item{lcYear2}{Column name (string) for year 2.}

\item{type}{Either "long" or "short".}
}
\value{
Character vector describing land cover transition.
}
\description{
Detect forest change between two land cover columns.
}
