% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc_eng.R
\name{lulc_eng}
\alias{lulc_eng}
\title{lulc_eng}
\usage{
lulc_eng(varName)
}
\arguments{
\item{varName}{Numeric, character, or factor vector of MoEF land cover codes.}
}
\value{
Character vector of land cover classes in English.
}
\description{
Recode MoEF land cover numeric codes into English class names.
}
