% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc_lubis.R
\name{lulc_lubis}
\alias{lulc_lubis}
\title{lulc_lubis}
\usage{
lulc_lubis(varName, type = c("long", "short"))
}
\arguments{
\item{varName}{Character or factor vector of land cover classes.}

\item{type}{Either "long" or "short".}
}
\value{
Recategorised land cover classes.
}
\description{
Recategorise land cover classes based on Lubis (2020).
Industrial plantation forest is classified as plantation (non-forest).
}
