% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc_moef.R
\name{lulc_moef}
\alias{lulc_moef}
\title{lulc_moef}
\usage{
lulc_moef(varName)
}
\arguments{
\item{varName}{Character or factor vector of land cover classes.}
}
\value{
Character vector of aggregated land cover categories.
}
\description{
Recategorise land cover classes based on MoEF regulation.
}
