% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulc_moef_codes.R
\name{lulc_moef_codes}
\alias{lulc_moef_codes}
\title{lulc_moef_codes}
\usage{
lulc_moef_codes(varName)
}
\arguments{
\item{varName}{A character vector of land use classes.}
}
\value{
Numeric vector of MoEF land cover codes.
}
\description{
Convert LULC description to official MoEF codes.
}
