% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vec.R
\name{scale_vec}
\alias{scale_vec}
\title{Scale a numeric vector (mean = 0, sd = 1)}
\usage{
scale_vec(x)
}
\arguments{
\item{x}{Numeric vector yang ingin di-scale.}
}
\value{
Vektor yang sudah distandardisasi.
}
\description{
Scale a numeric vector (mean = 0, sd = 1)
}
