% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_patrol_data.R
\name{sim_patrol_data}
\alias{sim_patrol_data}
\title{Sampling titik acak di sepanjang track dan memberi label spesies berdasarkan aturan ekologis}
\usage{
sim_patrol_data(
  track,
  dem,
  canopy_height,
  river_dist,
  n_points = 500,
  rules = data.frame(species = c("Harimau", "Gajah", "Rusa", "Babi hutan"), elev_max =
    c(2000, 800, 1500, 1000), canopy_min = c(20, 10, 15, 5), river_maxdist = c(2000,
    1000, 1500, 800), n = c(25, 50, 47, 60))
)
}
\arguments{
\item{track}{sf LINESTRING. Jalur survei atau lintasan.}

\item{dem}{SpatRaster elevasi (meter di atas permukaan laut).}

\item{canopy_height}{SpatRaster tinggi kanopi (meter).}

\item{river_dist}{SpatRaster jarak dari sungai (meter).}

\item{n_points}{integer, jumlah titik acak di sepanjang track.}

\item{rules}{data.frame berisi aturan spesies
(kolom wajib: species, elev_max, canopy_min, river_maxdist, n)}
}
\value{
sf POINT dengan kolom species, elev_m, canopy_m, dist_river, dan atribut tambahan.
}
\description{
Sampling titik acak di sepanjang track dan memberi label spesies berdasarkan aturan ekologis
}
