% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale_vec.R
\name{unscale_vec}
\alias{unscale_vec}
\title{Unscale a numeric vector using original mean and sd}
\usage{
unscale_vec(scaled, original)
}
\arguments{
\item{scaled}{Vektor yang sudah di-scale.}

\item{original}{Vektor referensi untuk menghitung mean dan sd.}
}
\value{
Vektor nilai asli (unscaled).
}
\description{
Unscale a numeric vector using original mean and sd
}
