% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2area.R
\name{dist2area}
\alias{dist2area}
\title{Calculate signed Euclidean distance from polygon boundary and interior}
\usage{
dist2area(polyData, rasterTemp, as_Raster = TRUE, rasRes = 1000)
}
\arguments{
\item{polyData}{An sf or SpatVector object representing the target polygon(s)}

\item{rasterTemp}{A template SpatRaster to define resolution and extent}

\item{as_Raster}{Logical; if TRUE, returns SpatRaster output (default = TRUE)}

\item{rasRes}{Numeric; desired raster resolution (default = 1000 meters)}
}
\value{
A SpatRaster object of signed Euclidean distances
}
\description{
This function computes a raster of signed distances relative to a polygon area,
where positive values represent distance \emph{outside} the polygon, and negative values
represent distance \emph{inside} the polygon. The result can be useful for modeling
habitat-edge effects or environmental gradients.
}
